Add 2024/06.2 TypeScript solution
The `console.log`s are commented out here by default. Without this, when
running `bun test` the methods run twice and it takes roughly twice as
long

Not a fan of a fairly nondescript `setup` function, but at least it has
a clear return type

I also moved the `visitedCoordinates` object onto the guard - it felt
like it belonged here, since the guard is the one who visits. This
object is updated to keep track of orientation, which is then used to
check for looping

Spent a long time wondering why new objects created in a loop were
updating old ones, having forgotten about the shallow vs deep copy
particulars with JavaScript objects:
- https://stackoverflow.com/a/68269665/4002016
- https://developer.mozilla.org/en-US/docs/Glossary/Deep_copy