Optimise 2024/06.2
Credit to @richardTowers for this one. I hadn't considered that obstacle
candidates can only include those cells visited in the original path.
Filtering by this cuts the part two runtime down from ~28s to ~4-5s

Since this allows the `setup()` part to be simplified, I opted against
separating this out into a solution two. In other cases where I've
provided two solutions, they tend to be more meaningfully distinct in
approach, whereas here I'm retaining the approach (so no real big O
change) but substantially reducing the N (which is impactful here). This
also meant reusing a slightly modified bit of part one, which is
therefore abstracted, so it's a refinement of the original approach(es)

https://github.com/richardTowers/advent-of-code/blob/main/2024/06/main.rb#L55