Add 2024/07.1 TypeScript solution
This was tricky. I learned a bit about permutations and things like
Heap's algorithm. I'm not sure if that's what I've implemented here, but
it works by iterating over the operators and keeping one element fixed,
then two, then three etc until its gone through all possible
permutations or found a match. This is achieved through recursion

Part one takes ~4s to execute, which feels okay given the number of
possible permutations that need to be checked. Though perhaps there's a
smarter way to do this and/or skip windows of permutations?

Almost got caught out at the end by the left to right execution, after
trying to build an expression string containing all the operators and
operands in one

I'm curious if there's a nicer way to implement the expression
evaluation than building expression strings and `eval`ing them, since
that wouldn't be very safe with dynamic/user-generated data. A switch/if
statement with cases for all the operators is an option, but it doesn't
feel very scalable, though perhaps limitless scaling (i.e. with an
arbitrary number of non-predefined operators) would be insecure in any
case