Optimise 2024/07.1 TypeScript solution
As I hinted at in the previous commit, I wondered about using a switch
statement instead of the `eval` call. This isn't very scalable, but
perhaps it shouldn't be. Turns out it is much more performant (at least
in TypeScript/JavaScript): before it took a little over 4s, now it takes
around 50ms or 0.05s