Add 2024/04.1 C#/.NET solution
In noticed that I used lists instead of arrays in the day 1 C#/.NET
solutions. In day two, `originalAndSubArrays` used arrays but only one
was modified, and that was reassigned with the original array spread
into a new array, instead of with pushing (via `Add` or `Append`). Day
three didn't have any arrays or lists

I encountered some challenges in growing arrays here, so I went back to
using lists for this purpose, which seems to popular:
https://stackoverflow.com/a/434765/4002016. I did however convert back
to an array once the size is fixed. Not sure if this is advisable or
whether there are performance gains from using an array or losses from
the conversion

C#/.NET will raise an out of bounds error if trying to access an array
element that doesn't exist by index, so in the `while` loop in
`rowsToColumns` the parenthetical current or default expression in
`columns[currentColumnIndex] = (columns[currentColumnIndex] || "") +
currentCharacter;` from the TypeScript solution won't work here. Instead
I checked the index against the length and added the default if it would
be out of bounds (before the statement that adds the current character).
Perhaps there's a cleaner/more succinct way of doing this?