Add opinionated C#/.NET formatter CSharpier
Before this, the main/vanilla C# extension was set as the default
formatter, but it didn't feel opinionated enough. It would retain empty
lines at the top of blocks and just not enforce a great deal of
consistency. I looked for an alternative and found CSharpier, which
appears to be inspired by Prettier

I installed the CSharpier VSCode extension and added the CSharpier
'tool' to the repo (perhaps independent of the 'solution') by running
`dotnet tool install csharpier --create-manifest-if-needed` in the
project root. I also updated my VSCode settings to format C# files on
save and use CSharpier as the default C# formatter