Add C#/.NET test suite and refactor 04 solution
I added a `tests` folder in the C# solutions folder, then ran `dotnet
new xunit` inside the new folder to create a test suite. I added this to
the solution with `dotnet sln add 2024/solutions/csharp/tests` in the
repository root

I added the a reference to the day 4 project to the tests project (as
documented in the C# README changes)

I then refactored the actual day 4 program into a class to enable unit
testing

I had some issues with getting the right file path for the input when
running the test suite - it was running relative to the test suite's
executable directory, meaning the relative path to the input went to a
different place. I tried a few things to make the path be relative to
the program class, but in the end I got around it by optionally taking
a path to the input file when constructing the class