Add 2024/09.2 TypeScript solution
It would be nice to break down the function here, but the algorithm
feels okay. It's very similar to part one, but just different enough to
not feel like it warranted abstracting anything common out (making
things similar enough might have actually reduced performance)

It took me an unfortunate amount of time to debug an issue where my
array was growing during the defragmentation stage. I was looking for
suitable spaces for move candidates in the whole array, rather than just
on the left side of a given move candidate

Switching to `console.table` instead of `console.log` really helped with
the debugging: seeing the index and all data of each file or empty space
object on a single line was very useful

Speaking of useful, so too was having the prompt browser in my terminal.
I ended up side-by-siding the prompt and either logging output or a
REPL, which allowed me to compared results of various steps with the
examples and values in the prompt

For both parts today, the general algorithm came quite naturally, and
they seem to be pretty performant (~3ms(!) for part one and a little
over 300ms for part two) but I made some pretty simple errors. Need more
sleep