Add 2024/11.2 TypeScript solution
I tried using multi-threading with promises and `Promise.all` over each
of the numbers, but that didn't make things fast enough at a depth of
75. I guess this is an example of where the big O really matters, not
the N, so dividing the N by 6 isn't going to help all that much

After that, I tried a few things, and I knew I wanted to go depth first
through blinks, memoising where each stone gets to at a given depth.
I played around with a few loops, but I knew I should probably use
recursion: without recursion, it would be a challenge to keep track of
the blinks since every ancestor node when diving down. In the end I
looked for some hints to help think through the structure of the
recursive part, and while I was slowly getting somewhere close,
`calculateStoneCountByBlinkCount` is broadly based on:
https://www.reddit.com/r/adventofcode/comments/1hbo9m6/comment/m1i35km