Add 2024/12.2 TypeScript solution
I considered walking the perimeter and counting every turn until
reaching the same coordinate with the same orientation, which would be
a bit of a call back to day 6, but I decided to try something different
and maybe a little more mathematical?

For each coordinate, it tracks the sides that don't neighbour another
plot in the region (using a Set). Then for each orientation and x combo,
it then filters the tracked coordinate/orientation objects. It takes
their y values, sorts them, and counts the contiguous groups of
integers. It then does the same with orientation and y combos, for each
counting their contiguous y integer groups

I think technically this has more steps than would the other approach,
where you just count the rotations, going through a maximum of the
perimeter cells * the directions. Here you go through each direction for
each coordinate in the region, and then have to do the additional
contiguous groups calculation at the end on each orientation for each
row and each column. But regardless I kind of like this approach, and
it takes around 65ms in my environment for this N count. Building a
rotating walker/checker hybrid could easily result in a monster

I almost had the right answer a few hours ago, but I hadn't realised or
had forgotten that I need to give `sort` a compare function even if I
just want a basic high to low or low to high sort... so when there were
11+ rows or columns it didn't get the right number of contiguous groups
because the numbers were improperly sorted. Of course none of the
examples have that many rows or columns, so it worked perfectly on them