Add 2024/13.1 TypeScript solution
I'm kinda surprised this works. I would've thought I'd be missing
instances where it's possible to reach the target with more Bs than As,
but cheaper with more As than Bs (if the A movements are on average more
than three times the B movements). Maybe the data are such that this
scenario isn't possible, and this seemingly naïve approach works

Also surprised that this is pretty performant given there are 320
machines that could each iterate through 10201 (0-100 a x 0-100 b)
combinations. There are of course early exits, but still...

First time using a class for AoC 2024