Add 2024/13.2 TypeScript solution
This also includes a divide and conquer version of part one and renames
the original solution to make it clear it's brute force

It's clear that there's only one good combination in each machine. I
thought that the data were handcrafted to ensure that, but I think it's
probably because for every a and b pair, if a has a higher x, b has a
higher y, so naturally there can only be one magic midpoint before one
side diverges

Working out the proximity calculation and then the divide and conquer
algorithm was a challenge here. And I spent a little time confused about
why I was getting a fractional count in the divide and conquer approach,
but I'd just forgotten to do a `% 1` check on counts with a proximity of
1