Add 2024/14.2 TypeScript solution
I'm not particularly fond of these solutions. They feel like exploits
based on characteristics of the data rather than solid algorithms. I
don't think there's any particular reason that these work other than the
data happening to be generated in a particular way

I didn't really get the prompt at first, so I hopped over to Reddit to
see if there were any threads explaining the problem. I stumbled on a
few things:
- printed maps showing what the tree should look like
- mention that there are no overlapping robots during the easter egg
  second
- mention that the easter egg second might be the only one with a
  certain number of a adjacent robots (in a line or a square or a
  contiguous region)
- mention of noticing that a near Christmas tree appeared at certain
  intervals, and the complete tree was on this interval
- mention that the configuration of robots must repeat every width *
  length seconds, since the robots loop around. Their x positions will
  loop every width seconds, and their y positions every height seconds,
  so to get them both to loop you need to go at most width * height
  seconds

Te second and third points feel like quirks of the data, though I
suppose you could intuit that looking for a substantial number of
adjacent robots would be useful

I implemented a couple of solutions along those lines. My basic
implementation of the adjacency check is pretty slow, but having seen
the interval hint and example trees, I had a think about optimisations.
I decided to go through a whole loop of permutations (width * height
seconds as mentioned above) recording the most robots in a column or row
for each, then check which second had the most. It turns out that the
second with the easter egg is a multiple of this second (or its index?).
The maximum robots in a line check was less expensive than the adjacency
check (in my implementation), so I did the former and then checked every
second that was a multiple of that second, which cut the time down to
close to (but still slightly slower than) the overlap checking solution

The fact that the second with the most robots in a single line is a
factor/divisor of the easter egg second - rather than another multiple
of a shared factor/divisor, for instance, or just completely unrelated
- again feels like a quirk of the data. Likely intentional, or an
artefact of the data generation method, but I'm not sure how you're
supposed to find this pattern without lucky guesswork