Add 2024/17.2 TypeScript solution
Can't take full credit for this one. I was trying to spot some patterns
in the numbers that were close, and then tried looking for 10-20 numbers
that produced a match for the first digit in the program, then taking
the smallest diff between any two of these digits, thinking it would be
the lowest common denominator. I then upped the increment to that,
restarted with the first match, and looked for a match for the first two
digits, and so on. This only got up to 12-digit matches, and it turns
out that while some diffs are multiples of other diffs at a given N of
digits, not all diffs are multiples, at least with matching from the
front

I wondered about reversing the instructions, but thought that you
wouldn't be able to precisely reverse a bitwise XOR, since presumably
multiple inputs could produce the same output, and in general it felt
like there would be too many unknowns to reverse things

I ended up having a browse through Reddit and came across a few
solutions that involved matching right to left and multiplying by 8.
Both of these had crossed my mind (I'd wondered about fixing one output
digit at a time when playing around with 14-15-digit numbers) but I
hadn't thought through any specifics

This finds the first number that produces an output with the last digit
of the program, then multiplies by eight and looks for the first number
that produces an output with the last two digits of the program, and so
on until all digits match. Credit to @famendola1 for the main algorithm
here:
- https://www.reddit.com/r/adventofcode/comments/1hg38ah/comment/m2og94d
-
  https://github.com/famendola1/aoc-2024/blob/master/src/advent_of_code/day17.clj

Without using BigInts, this will stop matching at 15 out of 16 digits

This algorithm is a little more complex but had a similar multiply by
eight step:
- https://www.reddit.com/r/adventofcode/comments/1hg38ah/comment/m2v7vmg
- https://github.com/goggle/AdventOfCode2024.jl/blob/main/src/day17.jl