Read/write inputs/prompts to/from separate repo
Inputs and prompts aren't meant to be published outside of
adventofcode.com. I've created a separate and private repo to store my
copies of these rather than bundling them inside this public solutions
repo. This introduces an environment variable which should be set to the
path of the 'assets' repo:
- from which all solutions should now get the input (all old solutions
  and tests should still work if I've tested everything thoroughly
  enough)
- to which (private) copies of newly-encountered inputs and prompts will
  be written by the `script/download.zsh`